Add basic setup info for virtualenv
I've put in the subfolder name here as venv because Heroku does it that
way (and assumes that all projects will be structured like that). Do we
want to stick with that? We can put it in a different folder for now or
in the root folder (which I think some of the existing gitignore
assumes), but that would make changing it around later if we do use
Heroku bothersome.

Note that I haven't yet looked up if any of the main Heroku competitors
assume the same folder placement.