add route and model for letsencrypt (ACME)
Let's Encrypt (https://letsencrypt.org/) is an SSL provider. To issue
certificates, it requires a customer's website to respond to a request
issued to an endpoint negotiated out of band, and expect a response
agreed upon at the same time as the endpoint.

They provide a script to automatically do this negotiation and response,
and the expected workflow is to run this script with root credentials so
that it can also automatically request and renew your SSL certificate,
as well as automatically reconfigure your httpd to use it.

That doesn't work out so well in a Heroku context. However, there are
more manual options. According to [1], letsencrypt expects its challenge
to be present at `/.well-known/acme-challenge/:id`, and in the manual
workflow gives you time to configure your webserver to respond
appropriately. In our case, that would be adding the challenge id and
expected response as a row in this newly-introduced table.

Once that challenge/response interaction is complete, the CA will sign
the certificate request, and the certificate can be uploaded to Heroku.

[1] http://collectiveidea.com/blog/archives/2016/01/12/lets-encrypt-with-a-rails-app-on-heroku/

Issue #144 Set up HTTPS